Split up core/util/serialization
