+++ bokehjs/src/lib/core/types.ts
@@ -2,6 +2,8 @@ export type uint8  = number
 export type uint16 = number
 export type uint32 = number
 
+export type ByteOrder = "little" | "big"
+
 export type ID = string
 
 export type Color = string
