+++ bokehjs/src/lib/core/ui_events.ts
@@ -7,7 +7,7 @@ import * as events from "./bokeh_events"
 import {getDeltaY} from "./util/wheel"
 import {reversed, is_empty} from "./util/array"
 import {isString} from "./util/types"
-import {is_mobile} from "./util/compat"
+import {is_mobile} from "./util/platform"
 import {PlotView} from "../models/plots/plot"
 import {ToolView} from "../models/tools/tool"
 import {ContextMenu} from "./util/menus"
