+++ bokehjs/src/lib/core/util/platform.ts
@@ -1,3 +1,5 @@
+import {ByteOrder} from "../types"
+
 export const is_ie = (() => {
   const ua = typeof navigator !== "undefined" ? navigator.userAgent : ""
   return ua.indexOf('MSIE') >= 0 || ua.indexOf('Trident') > 0 || ua.indexOf('Edge') > 0
@@ -19,3 +21,5 @@ export const is_little_endian = (() => {
   }
   return little_endian
 })()
+
+export const BYTE_ORDER: ByteOrder = is_little_endian ? "little" : "big"
