+++ bokehjs/src/lib/core/util/serialization.ts
@@ -1,67 +1,9 @@
-import {ID} from "../types"
+import {ID, ByteOrder} from "../types"
 import {isPlainObject} from "./types"
-import {is_little_endian} from "./compat"
 import {DataType, NDArray} from "./ndarray"
 import * as ndarray from "./ndarray"
-
-export function buffer_to_base64(buffer: ArrayBuffer): string {
-  const bytes = new Uint8Array(buffer)
-  const chars = Array.from(bytes).map((b) => String.fromCharCode(b))
-  return btoa(chars.join(""))
-}
-
-export function base64_to_buffer(base64: string): ArrayBuffer {
-  const binary_string = atob(base64)
-  const len = binary_string.length
-  const bytes = new Uint8Array(len)
-  for (let i = 0, end = len; i < end; i++) {
-    bytes[i] = binary_string.charCodeAt(i)
-  }
-  return bytes.buffer
-}
-
-export type ByteOrder = "little" | "big"
-
-export const BYTE_ORDER: ByteOrder = is_little_endian ? "little" : "big"
-
-export function swap16(a: Int16Array | Uint16Array): void {
-  const x = new Uint8Array(a.buffer, a.byteOffset, a.length * 2)
-  for (let i = 0, end = x.length; i < end; i += 2) {
-    const t = x[i]
-    x[i] = x[i + 1]
-    x[i + 1] = t
-  }
-}
-
-export function swap32(a: Int32Array | Uint32Array | Float32Array): void {
-  const x = new Uint8Array(a.buffer, a.byteOffset, a.length * 4)
-  for (let i = 0, end = x.length; i < end; i += 4) {
-    let t = x[i]
-    x[i] = x[i + 3]
-    x[i + 3] = t
-    t = x[i + 1]
-    x[i + 1] = x[i + 2]
-    x[i + 2] = t
-  }
-}
-
-export function swap64(a: Float64Array): void {
-  const x = new Uint8Array(a.buffer, a.byteOffset, a.length * 8)
-  for (let i = 0, end = x.length; i < end; i += 8) {
-    let t = x[i]
-    x[i] = x[i + 7]
-    x[i + 7] = t
-    t = x[i + 1]
-    x[i + 1] = x[i + 6]
-    x[i + 6] = t
-    t = x[i + 2]
-    x[i + 2] = x[i + 5]
-    x[i + 5] = t
-    t = x[i + 3]
-    x[i + 3] = x[i + 4]
-    x[i + 4] = t
-  }
-}
+import {BYTE_ORDER} from "./platform"
+import {swap, base64_to_buffer, buffer_to_base64} from "./buffer"
 
 export type Shape = number[]
 
@@ -113,17 +55,7 @@ export function decode_NDArray(ref: BufferRef | NDArrayRef, buffers: Buffers): N
   })()
 
   if (order !== BYTE_ORDER) {
-    switch (array.BYTES_PER_ELEMENT) {
-      case 2:
-        swap16(array as Int16Array | Uint16Array)
-        break
-      case 4:
-        swap32(array as Int32Array | Uint32Array | Float32Array)
-        break
-      case 8:
-        swap64(array as Float64Array)
-        break
-    }
+    swap(array)
   }
 
   return array
