+++ bokehjs/src/lib/models/tools/gestures/wheel_zoom_tool.ts
@@ -3,7 +3,7 @@ import {scale_range} from "core/util/zoom"
 import * as p from "core/properties"
 import {PinchEvent, ScrollEvent} from "core/ui_events"
 import {Dimensions} from "core/enums"
-import {is_mobile} from "core/util/compat"
+import {is_mobile} from "core/util/platform"
 import {bk_tool_icon_wheel_zoom} from "styles/icons"
 
 export class WheelZoomToolView extends GestureToolView {
